unused file removal..

+++ src/traps/smasher.cc
@@ -1,4 +1,4 @@
-//  $Id: smasher.cc,v 1.21 2001/04/27 20:44:38 grumbel Exp $
+//  $Id: smasher.cc,v 1.22 2001/08/04 11:53:59 grumbel Exp $
 //
 //  Pingus - A free Lemmings clone
 //  Copyright (C) 1999 Ingo Ruhnke <grumbel@gmx.de>
@@ -53,7 +53,7 @@ Smasher::update(float delta)
 	      // SMASH!!! The thing hitten earth and kills the pingus
 	      downwards = false;
 	      --count; 
-	      PingusSound::play_wav("tenton", 0.7);
+	      PingusSound::play_wav("tenton", 0.7f);
 	      
 	      for(int i=0; i < 20; i++)
 		{
@@ -98,13 +98,13 @@ Smasher::draw_colmap()
   std::cout << "Drawing colmap entry" << std::endl;
 
   CL_Surface prov(PingusResource::load_surface("Traps/smasher_cmap", "traps"));
-  world->get_colmap()->put(prov, pos.x, pos.y, GroundpieceData::SOLID);
+  world->get_colmap()->put(prov, (int) pos.x, (int) pos.y, GroundpieceData::SOLID);
 }
 
 void 
 Smasher::draw_offset(int x, int y, float s)
 {
-  surface.put_screen(pos.x + x, pos.y + y, count);
+  surface.put_screen(int(pos.x + x), int(pos.y + y), count);
 }
 
 void
